split filename generation into a dedicated class
