+++ aqua/core/regridder/regridder.py
@@ -1,26 +1,24 @@
 """New Regrid class independent from the Reader"""
 
 import os
-import re
 import shutil
 from tempfile import TemporaryDirectory
 
 import xarray as xr
 from smmregrid import CdoGenerate, GridInspector
 from smmregrid import Regridder as SMMRegridder
-from smmregrid.util import check_gridfile
 
 from aqua.core.default import (
     DEFAULT_DIMENSION,
     DEFAULT_DIMENSION_MASK,
     DEFAULT_REGRID_METHOD,
-    DEFAULT_WEIGHTS_AREAS_PARAMETERS,
 )
 from aqua.core.logger import log_configure
 from aqua.core.util import to_list
 
-from .griddicthandler import GridDictHandler
-from .regridder_util import check_existing_file
+from .filename_handler import FilenameHandler
+from .grid_dict_handler import GridDictHandler
+from .regridder_util import check_existing_file, get_grid_path
 
 # please notice: check_gridfile is a function from smmregrid.util
 # to check and if a grid is a cdo grid,
@@ -87,6 +85,8 @@ class Regridder:
 
         self.logger.debug("Normalized grid dictionary: %s", self.src_grid_dict)
         self.logger.debug("Normalized grid path: %s", self.src_grid_path)
+
+        self.filenames = FilenameHandler(self.cfg_grid_dict, self.src_grid_name, self.loglevel)
 
         # this not used but can be shipped back to the reader
         self.src_horizontal_dims = self.src_grid_dict.get("space_coord", None)
@@ -253,7 +253,7 @@ class Regridder:
 
             xr.Dataset: The computed grid area.
         """
-        area_filename = self._area_filename(grid_name if grid_name else None, regridder_metadata)
+        area_filename = self.filenames.area_filename(grid_name if grid_name else None, self.src_grid_path, regridder_metadata)
         area_type = "target" if grid_name else "source"
 
         # if file exists, load it
@@ -293,8 +293,8 @@ class Regridder:
 
         self.logger.info("Generating %s area for %s", area_type, grid_name)
 
-        source_grid = self._get_grid_path(grid_dict.get("path")) if area_type == "source" else None
-        target_grid = self._get_grid_path(grid_dict.get("path")) if area_type == "target" else None
+        source_grid = get_grid_path(grid_dict.get("path")) if area_type == "source" else None
+        target_grid = get_grid_path(grid_dict.get("path")) if area_type == "target" else None
 
         return CdoGenerate(
             source_grid=source_grid,
@@ -340,7 +340,9 @@ class Regridder:
             # define the vertical coordinate in the smmregrid world
             smm_mask_dim = None if mask_dim in [DEFAULT_DIMENSION, DEFAULT_DIMENSION_MASK] else mask_dim
 
-            weights_filename = self._weights_filename(tgt_grid_name, regrid_method, mask_dim, regridder_metadata)
+            weights_filename = self.filenames.weights_filename(
+                tgt_grid_name, self.src_grid_path, regrid_method, mask_dim, regridder_metadata
+            )
 
             # check if weights already exist, if not, generate them
             if rebuild or not check_existing_file(weights_filename):
@@ -357,7 +359,7 @@ class Regridder:
                 # and reduce the dimensionality of the input data.
                 generator = CdoGenerate(
                     source_grid=self.src_grid_path[mask_dim],
-                    target_grid=self._get_grid_path(tgt_grid_dict.get("path")),
+                    target_grid=get_grid_path(tgt_grid_dict.get("path")),
                     cdo_extra=cdo_extra,
                     cdo_options=cdo_options,
                     cdo=self.cdo,
@@ -400,100 +402,6 @@ class Regridder:
                 mask_dim=smm_mask_dim,
                 loglevel=self.loglevel,
             )
-
-    def _area_filename(self, tgt_grid_name, regridder_metadata):
-        """
-        Generate the area filename with intelligent fallback for path-based backends.
-
-        Args:
-            tgt_grid_name (str): The destination grid name (None for source grid).
-            regridder_metadata (RegridderMetadata): Metadata for filename generation.
-
-        Returns:
-            str: The area filename
-        """
-
-        area_dict = self.cfg_grid_dict.get("areas")
-
-        if not area_dict:
-            self.logger.warning("Areas block not found in the configuration file, using fallback naming scheme.")
-            if tgt_grid_name:
-                return f"cell_area_{tgt_grid_name}.nc"
-            return self._fallback_filename(regridder_metadata, kind="area")
-
-        # Target grid: always use grid template
-        if tgt_grid_name:
-            filename = area_dict["template_grid"].format(grid=tgt_grid_name)
-            self.logger.debug("Using grid-based template for target grid. Filename: %s", filename)
-        # Source grid: check if it's a file or data
-        else:
-            if check_gridfile(self._get_grid_path(self.src_grid_path)) != "xarray":
-                # Grid file: use grid name
-                filename = area_dict["template_grid"].format(grid=self.src_grid_name)
-                self.logger.debug("Using grid-based template for source grid. Filename: %s", filename)
-            else:
-                # Grid from data: use metadata
-                filename = self._resolve_source_filename(area_dict, regridder_metadata, kind="area")
-
-        filename = self._insert_metadata_params(filename, regridder_metadata)
-        filename = self._filename_prepend_path(filename, kind="areas")
-        return filename
-
-    def _weights_filename(self, tgt_grid_name, regrid_method, mask_dim, regridder_metadata):
-        """
-        Generate the weights filename with intelligent fallback for path-based backends.
-
-        Args:
-            tgt_grid_name (str): The destination grid name.
-            regrid_method (str): The regrid method.
-            mask_dim (str): The vertical dimension.
-            regridder_metadata (RegridderMetadata): Metadata for filename generation.
-
-        Returns:
-            str: The weights filename
-        """
-
-        levname = mask_dim if mask_dim in [DEFAULT_DIMENSION, DEFAULT_DIMENSION_MASK] else f"3d-{mask_dim}"
-
-        weights_dict = self.cfg_grid_dict.get("weights")
-
-        if not weights_dict:
-            self.logger.warning("Weights block not found in the configuration file, using fallback naming scheme.")
-            return f"weights_{tgt_grid_name}_{regrid_method}_l{levname}.nc"
-
-        # Check if grid is from file or data
-        if check_gridfile(self.src_grid_path[mask_dim]) != "xarray":
-            # Grid file: use grid-based template
-            filename = weights_dict["template_grid"].format(
-                sourcegrid=self.src_grid_name, method=regrid_method, targetgrid=tgt_grid_name, level=levname
-            )
-            self.logger.debug("Using grid-based template for weights. Filename: %s", filename)
-        else:
-            # Grid from data: use unified resolution logic
-            filename = self._resolve_source_filename(
-                weights_dict, regridder_metadata, kind="weights", method=regrid_method, targetgrid=tgt_grid_name, level=levname
-            )
-
-        filename = self._insert_metadata_params(filename, regridder_metadata)
-        filename = self._filename_prepend_path(filename, kind="weights")
-        return filename
-
-    def _filename_prepend_path(self, filename, kind="weights"):
-        """
-        Prepend the path to the filename with some fall back option
-        """
-        if not self.cfg_grid_dict.get("paths"):
-            self.logger.warning("Paths block not found in the configuration file, using present directory.")
-        else:
-            if not self.cfg_grid_dict["paths"].get(kind):
-                self.logger.warning("%s block not found in the paths block, using present directory.", kind)
-            else:
-                # if path does not exist, create it
-                if not os.path.exists(self.cfg_grid_dict["paths"][kind]):
-                    self.logger.warning("%s path in %s does not exist: creating!", kind, self.cfg_grid_dict["paths"][kind])
-                    os.makedirs(self.cfg_grid_dict["paths"][kind], exist_ok=True)
-                filename = os.path.join(self.cfg_grid_dict["paths"][kind], filename)
-        return filename
 
     def _expand_dims(self, data, mask_dims):
         """
@@ -602,138 +510,6 @@ class Regridder:
             raise ValueError("Data must be an xarray Dataset or DataArray.")
         return data
 
-    def _resolve_source_filename(self, config_dict, regridder_metadata, kind="area", **extra_params):
-        """Resolve filename for data-based grids with smart fallback.
-
-        Works for both areas and weights by accepting extra template parameters.
-
-        Args:
-            config_dict: Config section (areas or weights)
-            regridder_metadata: RegridderMetadata instance
-            kind: "area" or "weights" (for logging and fallback)
-            **extra_params: Extra params for template (e.g., method, targetgrid, level)
-
-        Returns:
-            Resolved filename string
-        """
-        if regridder_metadata and regridder_metadata.is_catalog_based():
-            # Catalog-based: use template_default
-            template = config_dict.get("template_default")
-            if not template:
-                raise ValueError(f"template_default missing in {kind} config")
-
-            self.logger.debug("Using catalog-based template for %s", kind)
-            # Merge metadata dict with extra params
-            template_vars = regridder_metadata.to_template_dict()
-            template_vars.update(extra_params)
-            return template.format(**template_vars)
-
-        elif regridder_metadata and regridder_metadata.is_path_based():
-            # Path-based: try template_path or use fallback
-            template_path = config_dict.get("template_path")
-
-            if template_path:
-                # Config has path-specific templa
-                self.logger.debug("Using path-based template for %s", kind)
-                path_id = regridder_metadata.get_path_identifier()
-                template_vars = {"path_id": path_id}
-                template_vars.update(extra_params)
-                return template_path.format(**template_vars)
-            else:
-                # No path template: use identifier fallback
-                self.logger.warning("No template_path in config, using fallback for %s", kind)
-                return self._fallback_filename(regridder_metadata, kind, **extra_params)
-
-        else:
-            # No metadata or invalid metadata
-            self.logger.warning("No valid regridder metadata, using fallback")
-            return self._fallback_filename(regridder_metadata, kind, **extra_params)
-
-    def _fallback_filename(self, regridder_metadata, kind="area", **extra_params):
-        """Generate fallback filename when templates unavailable.
-
-        Works for both areas and weights.
-
-        Args:
-            regridder_metadata: RegridderMetadata instance or None
-            kind: "area" or "weights" (for filename generation)
-            **extra_params: Extra params (e.g., method, targetgrid, level for weights)
-
-        Returns:
-            Fallback filename string
-        """
-        if kind == "area":
-            # Area fallback
-            if self.src_grid_name:
-                return f"cell_area_{self.src_grid_name}.nc"
-            elif regridder_metadata and regridder_metadata.path:
-                path_id = regridder_metadata.get_path_identifier()
-                return f"cell_area_path_{path_id}.nc"
-            else:
-                self.logger.warning(
-                    "No regridder_metadata or src_grid_name available for area filename generation. "
-                    "Using 'cell_area_unknown.nc'. Consider providing RegridderMetadata."
-                )
-                return "cell_area_unknown.nc"
-        else:
-            # Weights fallback
-            method = extra_params.get("method", "unknown")
-            targetgrid = extra_params.get("targetgrid", "unknown")
-            level = extra_params.get("level", "2d")
-
-            if regridder_metadata and regridder_metadata.path:
-                path_id = regridder_metadata.get_path_identifier()
-                return f"weights_path_{path_id}_{method}_{targetgrid}_l{level}.nc"
-            elif self.src_grid_name:
-                return f"weights_{self.src_grid_name}_{method}_{targetgrid}_l{level}.nc"
-            else:
-                self.logger.warning(
-                    "No regridder_metadata or src_grid_name available for weights filename generation. "
-                    "Using 'weights_unknown_%s_%s_l%s.nc'. Consider providing RegridderMetadata.",
-                    method,
-                    targetgrid,
-                    level,
-                )
-                return f"weights_unknown_{method}_{targetgrid}_l{level}.nc"
-
-    def _insert_metadata_params(self, filename, regridder_metadata):
-        """Insert extra parameters from DEFAULT_WEIGHTS_AREAS_PARAMETERS into filename.
-
-        Replaces _insert_kwargs() with RegridderMetadata-aware version.
-
-        Args:
-            filename: The filename to modify
-            regridder_metadata: RegridderMetadata instance
-
-        Returns:
-            Modified filename with parameters inserted
-        """
-        if not regridder_metadata:
-            return filename
-
-        for param in DEFAULT_WEIGHTS_AREAS_PARAMETERS:
-            value = getattr(regridder_metadata, param, None)
-            if value is not None:
-                filename = re.sub(r"\.nc", f"_{param}{value}.nc", filename)
-
-        return filename
-
-    @staticmethod
-    def _insert_kwargs(filename, reader_kwargs):
-        """
-        Insert the DEFAULT_WEIGHTS_AREAS_PARAMETERS in the filename template.
-
-        DEPRECATED: Use _insert_metadata_params() instead.
-        Kept for backward compatibility only.
-        """
-        # add the kwargs naming in the template file
-        if isinstance(reader_kwargs, dict):
-            for parameter in DEFAULT_WEIGHTS_AREAS_PARAMETERS:
-                if parameter in reader_kwargs:
-                    filename = re.sub(r"\.nc", "_" + parameter + str(reader_kwargs[parameter]) + r"\g<0>", filename)
-
-        return filename
-
     @staticmethod
     def configure_masked_fields(src_grid_dict):
         """
@@ -755,12 +531,3 @@ class Regridder:
         masked_attrs = {k: v for k, v in masked_info.items() if k != "vars"} or None
 
         return masked_attrs, masked_vars
-
-    @staticmethod
-    def _get_grid_path(grid_path):
-        """
-        Get the grid path from the grid dictionary.
-        This looks for `DEFAULT_DIMENSION` key,
-        otherwise takes the first available value in the dict.
-        """
-        return grid_path.get(DEFAULT_DIMENSION, next(iter(grid_path.values()), None))
