+++ aqua/core/regridder/regridder_util.py
@@ -2,6 +2,8 @@
 
 import os
 
+from aqua.core.default import DEFAULT_DIMENSION
+
 
 def check_existing_file(filename):
     """
@@ -9,3 +11,8 @@ def check_existing_file(filename):
     Return true if the file has some records.
     """
     return os.path.exists(filename) and os.path.getsize(filename) > 0
+
+
+def get_grid_path(grid_path):
+    """Get the grid path, looking for DEFAULT_DIMENSION or falling back to the first value."""
+    return grid_path.get(DEFAULT_DIMENSION, next(iter(grid_path.values()), None))
